backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioredis==2.0.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
async-timeout==5.0.1
asyncinit==0.2.4
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
docker==7.1.0
ecdsa==0.19.1
exceptiongroup==1.3.1
expiringdict==1.2.2
fastapi==0.128.0
fastapi-cache==0.1.0
fastapi-cache2==0.2.2
frozenlist==1.8.0
funcy==2.0
graphenelib==1.6.1
h11==0.16.0
hiredis==3.3.0
idna==3.11
multiavatar==1.0.3
multidict==6.7.0
packaging==25.0
pendulum==3.1.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pylibscrypt==2.0.0
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
scrypt==0.9.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
time-machine==3.2.0
toolz==1.1.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
viz-python-lib==1.1.0
websocket-client==1.9.0
websockets==15.0.1
pyeditorjs==1.5.1
yarl==1.22.0
